+++ packages/insomnia/src/ui/index.tsx
@@ -284,12 +284,12 @@ async function renderApp() {
                   {
                     path: 'collaborators',
                     loader: async (...args) =>
-                      (await import('./routes/collaborators')).collaboratorsListLoader(...args),
+                      (await import('./routes/$organizationId.collaborators')).collaboratorsListLoader(...args),
                   },
                   {
                     path: 'collaborators-search',
                     loader: async (...args) =>
-                      (await import('./routes/collaborators')).collaboratorSearchLoader(...args),
+                      (await import('./routes/$organizationId.collaborators')).collaboratorSearchLoader(...args),
                   },
                   {
                     path: 'invites',
@@ -298,12 +298,14 @@ async function renderApp() {
                         path: ':invitationId',
                         id: ':invitationId',
                         action: async (...args) =>
-                          (await import('./routes/collaborators')).updateInvitationRoleAction(...args),
+                          (await import('./routes/$organizationId.collaborators')).updateInvitationRoleAction(...args),
                         children: [
                           {
                             path: 'reinvite',
                             action: async (...args) =>
-                              (await import('./routes/collaborators')).reinviteCollaboratorAction(...args),
+                              (await import('./routes/$organizationId.collaborators')).reinviteCollaboratorAction(
+                                ...args,
+                              ),
                           },
                         ],
                       },
@@ -319,7 +321,7 @@ async function renderApp() {
                           {
                             path: 'roles',
                             action: async (...args) =>
-                              (await import('./routes/collaborators')).updateMemberRoleAction(...args),
+                              (await import('./routes/$organizationId.collaborators')).updateMemberRoleAction(...args),
                           },
                         ],
                       },
